requirement.txt
glob
tqdm
email.message
python-pptx
collections
collections.abc
csv
comtypes